Reorganization of the API * Make the raw API private * Separate simplified abstractions of each API in a separate file
